bot: refresRsvp if updating attendees from the event
This is the second half of the fix for messages not being sent
on updateCalendarEvents()

The problem was that refreshRsvp first sets hide=true because the
invite list is empty, as updateRsvp does not set this list.
addInvites and syncInvites modify that list, but then never called
refreshRsvp to set hide=false again.